Rename hugging-face-3-2-gptneo-text-generation(1).ipynb to Kagglehugging-face-3-2-gptneo-text-generation.ipynb
This version of this notebook was from Kaggle and actually ran. The version from Colab will freeze as the model is too large for Google Colab.   When running on Kaggle it may ask you to upgrade to Google AI notebooks, but you don't have to do that, just dismiss the prompt and let it finish running.